refactor: msw 제거
