+++ .env.local.example
@@ -2,9 +2,10 @@
 NEXT_PUBLIC_API_MOCKING="enable" 
 
 # 백엔드 api 주소를 결정합니다.
-NEXT_PUBLIC_BACK_API="http://localhost:8080"
+NEXT_PUBLIC_BACK_URL="http://localhost:8080"
 
 # 클라이언트 api 주소를 설정합니다.
-NEXT_PUBLIC_CLIENT_API="http://localhost:3000"
+NEXT_PUBLIC_CLIENT_URL="http://localhost:3000"
+
 # 서버 api 주소를 설정합니다.
-NEXT_PUBLIC_SERVER_API="http://localhost:3000"
+NEXT_PUBLIC_SERVER_URL="http://localhost:3000"
