+++ package.json
@@ -24,15 +24,16 @@
     }
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.17.15",
+    "@tanstack/react-query": "^5.17.19",
     "axios": "^1.6.5",
     "clsx": "^2.1.0",
-    "msw": "^2.1.4",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.49.3",
+    "run": "^1.5.0",
     "tailwind-merge": "^2.2.1",
+    "update": "^0.7.4",
     "zod": "^3.22.4",
     "zustand": "^4.5.0"
   },
