--- src/__mock__/handler/client/index.ts
@@ -1,5 +1,0 @@
-import { setupWorker } from "msw/browser"
-
-import tempClientHandler from "./temp-client-handler"
-
-export const clientWorker = setupWorker(...tempClientHandler)
