--- src/__mock__/handler/server/index.ts
@@ -1,5 +1,0 @@
-import { setupServer } from "msw/node"
-
-import tempServerHandler from "./temp-server-handler"
-
-export const serverWorker = setupServer(...tempServerHandler)
