+++ src/app/_component/temp-client/index.tsx
@@ -3,7 +3,7 @@
 import { useGetUsers } from "@/services/temp/hook"
 
 export default function TempClient() {
-  const { data: users } = useGetUsers()
-
-  return <div className="temp-btn">{JSON.stringify(users?.message)}</div>
+  const { data: user } = useGetUsers()
+  console.log(process.env.NEXT_PUBLIC_API_MOCKING)
+  return <div className="temp-btn">{JSON.stringify(user?.message)}</div>
 }
