+++ src/app/apis/temp/route.ts
@@ -1,0 +1,7 @@
+import { NextResponse } from "next/server"
+
+import { tempMock } from "@/seeds/temp-mock"
+
+export async function GET() {
+  return NextResponse.json(tempMock)
+}
