+++ src/components/provider/index.tsx
@@ -1,0 +1,16 @@
+"use client"
+
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import React, { PropsWithChildren } from "react"
+
+const Providers = ({ children }: PropsWithChildren) => {
+  const queryClient = new QueryClient({
+    defaultOptions: {
+      queries: { refetchOnWindowFocus: false, retry: false },
+    },
+  })
+
+  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+}
+
+export default Providers
