+++ src/constants/apis.ts
@@ -1,6 +1,8 @@
-export const BACK_URL = process.env.NEXT_PUBLIC_BACK_API
+export const BACK_URL = process.env.NEXT_PUBLIC_API_MOCKING === "enable" ? "/apis" : process.env.NEXT_PUBLIC_BACK_URL
 
-export const CLIENT_URL = process.env.NEXT_PUBLIC_CLIENT_API
+export const CLIENT_URL = process.env.NEXT_PUBLIC_CLIENT_URL
+
+export const SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL
 
 export const END_POINT = {
   temp: "/temp",
