--- src/libs/msw/mock-enable.ts
@@ -1,11 +1,0 @@
-async function mockEnable() {
-  if (typeof window === "undefined") {
-    const { serverWorker } = await import("@/__mock__/handler/server")
-    await serverWorker.listen({ onUnhandledRequest: "bypass" })
-  } else {
-    const { clientWorker } = await import("@/__mock__/handler/client")
-    await clientWorker.start({ onUnhandledRequest: "bypass" })
-  }
-}
-
-export default mockEnable
