+++ src/services/temp/api.ts
@@ -1,7 +1,6 @@
 import axios from "axios"
 
-import { END_POINT } from "@/constants/apis"
-import { mswInstance } from "@/libs/axios/instance"
+import { BACK_URL, END_POINT } from "@/constants/apis"
 
 type Temp = {
   message: string
@@ -10,19 +9,7 @@ type Temp = {
 const tempAPI = {
   getClientTemp: async () => {
     try {
-      const res = await mswInstance.get<Temp>(END_POINT.temp)
-      return res.data
-    } catch (err) {
-      if (err instanceof Error) {
-        throw new Error(err.message)
-      }
-    }
-  },
-
-  getServerTemp: async () => {
-    try {
-      // const res = await mswInstance.get<Temp>(END_POINT.temp)
-      const res = await axios("https://jsonplaceholder.typicode.com/todos/1")
+      const res = await axios.get<Temp>(BACK_URL + END_POINT.temp)
       return res.data
     } catch (err) {
       if (err instanceof Error) {
